redesign: data folder structure
