--- data/load_catalog/__init__.py
@@ -1,3 +1,0 @@
-# flake8: noqa
-
-from .load_catalog import load
