+++ models/load_model.py
@@ -6,13 +6,12 @@ from random import seed
 import numpy as np
 import pandas as pd
 import xgboost as xgb
-from matplotlib import pyplot as plt
 from sklearn.ensemble import RandomForestClassifier
 from sklearn.linear_model import LogisticRegression
 from sklearn.neural_network import MLPClassifier
 from sklearn.tree import DecisionTreeClassifier
 
-import data.catalog.loadData as loadData
+import data.load_data as loadData
 import tools.utils.memoize as utils
 from models.utils.implement_framework_model import (
     PyTorchLogisticRegression,
@@ -189,126 +188,3 @@ def loadModelForDataset(
         )
 
     return model_trained
-
-
-def scatterDataset(dataset_obj, classifier_obj, ax):
-    assert len(dataset_obj.getInputAttributeNames()) <= 3
-    X_train, X_test, y_train, y_test = dataset_obj.getTrainTestSplit()
-    X_train_numpy = X_train.to_numpy()
-    X_test_numpy = X_test.to_numpy()
-    y_train = y_train.to_numpy()
-    y_test = y_test.to_numpy()
-    number_of_samples_to_plot = min(200, X_train_numpy.shape[0], X_test_numpy.shape[0])
-    for idx in range(number_of_samples_to_plot):
-        color_train = "black" if y_train[idx] == 1 else "magenta"
-        color_test = "black" if y_test[idx] == 1 else "magenta"
-        if X_train.shape[1] == 2:
-            ax.scatter(
-                X_train_numpy[idx, 0],
-                X_train_numpy[idx, 1],
-                marker="s",
-                color=color_train,
-                alpha=0.2,
-                s=10,
-            )
-            ax.scatter(
-                X_test_numpy[idx, 0],
-                X_test_numpy[idx, 1],
-                marker="o",
-                color=color_test,
-                alpha=0.2,
-                s=15,
-            )
-        elif X_train.shape[1] == 3:
-            ax.scatter(
-                X_train_numpy[idx, 0],
-                X_train_numpy[idx, 1],
-                X_train_numpy[idx, 2],
-                marker="s",
-                color=color_train,
-                alpha=0.2,
-                s=10,
-            )
-            ax.scatter(
-                X_test_numpy[idx, 0],
-                X_test_numpy[idx, 1],
-                X_test_numpy[idx, 2],
-                marker="o",
-                color=color_test,
-                alpha=0.2,
-                s=15,
-            )
-
-
-def scatterDecisionBoundary(dataset_obj, classifier_obj, ax):
-    if len(dataset_obj.getInputAttributeNames()) == 2:
-        x_range = ax.get_xlim()[1] - ax.get_xlim()[0]
-        y_range = ax.get_ylim()[1] - ax.get_ylim()[0]
-        X = np.linspace(
-            ax.get_xlim()[0] - x_range / 10, ax.get_xlim()[1] + x_range / 10, 1000
-        )
-        Y = np.linspace(
-            ax.get_ylim()[0] - y_range / 10, ax.get_ylim()[1] + y_range / 10, 1000
-        )
-        X, Y = np.meshgrid(X, Y)
-        Xp = X.ravel()
-        Yp = Y.ravel()
-
-        # if normalized_fixed_model is False:
-        #   labels = classifier_obj.predict(np.c_[Xp, Yp])
-        # else:
-        #   Xp = (Xp - dataset_obj.attributes_kurz['x0'].lower_bound) / \
-        #        (dataset_obj.attributes_kurz['x0'].upper_bound - dataset_obj.attributes_kurz['x0'].lower_bound)
-        #   Yp = (Yp - dataset_obj.attributes_kurz['x1'].lower_bound) / \
-        #        (dataset_obj.attributes_kurz['x1'].upper_bound - dataset_obj.attributes_kurz['x1'].lower_bound)
-        #   labels = classifier_obj.predict(np.c_[Xp, Yp])
-        labels = classifier_obj.predict(np.c_[Xp, Yp])
-        Z = labels.reshape(X.shape)
-
-        cmap = plt.get_cmap("Paired")
-        ax.contourf(X, Y, Z, cmap=cmap, alpha=0.5)
-
-    elif len(dataset_obj.getInputAttributeNames()) == 3:
-        fixed_model_w = classifier_obj.coef_
-        fixed_model_b = classifier_obj.intercept_
-
-        x_range = ax.get_xlim()[1] - ax.get_xlim()[0]
-        y_range = ax.get_ylim()[1] - ax.get_ylim()[0]
-        X = np.linspace(
-            ax.get_xlim()[0] - x_range / 10, ax.get_xlim()[1] + x_range / 10, 10
-        )
-        Y = np.linspace(
-            ax.get_ylim()[0] - y_range / 10, ax.get_ylim()[1] + y_range / 10, 10
-        )
-        X, Y = np.meshgrid(X, Y)
-        Z = (
-            -(fixed_model_w[0][0] * X + fixed_model_w[0][1] * Y + fixed_model_b)
-            / fixed_model_w[0][2]
-        )
-
-
-def visualizeDatasetAndFixedModel(dataset_obj, classifier_obj, experiment_folder_name):
-    if not len(dataset_obj.getInputAttributeNames()) <= 3:
-        return
-
-    if len(dataset_obj.getInputAttributeNames()) == 2:
-        ax = plt.subplot()
-        ax.set_xlabel("x1")
-        ax.set_ylabel("x2")
-        ax.grid()
-    elif len(dataset_obj.getInputAttributeNames()) == 3:
-        ax = plt.subplot(1, 1, 1, projection="3d")
-        ax.set_xlabel("x1")
-        ax.set_ylabel("x2")
-        ax.set_zlabel("x3")
-        ax.view_init(elev=10, azim=-20)
-
-    scatterDataset(dataset_obj, classifier_obj, ax)
-    scatterDecisionBoundary(dataset_obj, classifier_obj, ax)
-
-    ax.set_title(f"{dataset_obj.dataset_name}")
-    ax.grid(True)
-
-    # plt.show()
-    plt.savefig(f"{experiment_folder_name}/_dataset_and_model.pdf")
-    plt.close()
